requirements.txt
streamlit
google-generativeai
pillow
opencv-python
reportlab
